add the connectivity location moos app

+++ src/pCommunicationAngle/AcousticPath.h
@@ -15,41 +15,41 @@ class AcousticPath
 
 
   //Set functions for initial params
-  void setSurfaceSoundSpeed(float c_0) {m_c_0=c_0;};
-  void setTimeInterval(float time_int){m_time_int=time_int;};
-  void setWaterDepth(float water_depth){m_water_depth=water_depth;};
-  void setSoundSpeedGrad(float gradient){m_gradient= gradient;};
+  void setSurfaceSoundSpeed(double c_0) {m_c_0=c_0;};
+  void setTimeInterval(double time_int){m_time_int=time_int;};
+  void setWaterDepth(double water_depth){m_water_depth=water_depth;};
+  void setSoundSpeedGrad(double gradient){m_gradient= gradient;};
   //Calculation functions for Acoustic Path
-  float calcProj_r(float x_rec,float y_rec, float x_src, float y_src);
-  float calcC(float z);
-  float calcMidpt(float r_1,float r_2);
-  float calcPerpSlope(float r_1,float z_1, float r_2, float z_2);
-  float calcPerpIntercept(float slope, float midpt_r,float midpt_z);
-  float calcCircCenter_r(float z_center, float int_slope, float int_b);
+  double calcProj_r(double x_rec,double y_rec, double x_src, double y_src);
+  double calcC(double z);
+  double calcMidpt(double r_1,double r_2);
+  double calcPerpSlope(double r_1,double z_1, double r_2, double z_2);
+  double calcPerpIntercept(double slope, double midpt_r,double midpt_z);
+  double calcCircCenter_r(double z_center, double int_slope, double int_b);
 
-  float calcCircCenter_z();
-  float calcRBisect(float r_1,float z_1,float r_center, float z_center);
-  float calcThetaSrc(float R,float z_src);
-  bool checkValidR(float R);
-  float calcValidR(float R_current);
-  float calcNewCircCenter_r(float z_rec, float r_rec,float R_new,float circ_z_center);
-  float calcPosOnCirc_r(float circ_z_center,float circ_r_center, float z_src, float R);
-  float convertRad2Degrees(float angle);
+  double calcCircCenter_z();
+  double calcRBisect(double r_1,double z_1,double r_center, double z_center);
+  double calcThetaSrc(double R,double z_src);
+  bool checkValidR(double R);
+  double calcValidR(double R_current);
+  double calcNewCircCenter_r(double z_rec, double r_rec,double R_new,double circ_z_center);
+  double calcPosOnCirc_r(double circ_z_center,double circ_r_center, double z_src, double R);
+  double convertRad2Degrees(double angle);
 
   //Calculation functions for Transmission Loss
-  float calcArcLength(float R, float theta_src, float theta_rec);
-  float calcThetaRec(float z_src, float z_rec, float theta_src);
-  float calc_r(float theta_src,float s, float R);
-  float calcJ(float s, float theta_src, float d_theta, float R);
-  float calcPfromArcLength(float z_src, float z_rec, float J, float theta_src);
-  float calcTransmissionLoss(float theta_src, float z_src, float z_rec, float R, float d_theta);
+  double calcArcLength(double R, double theta_src, double theta_rec);
+  double calcThetaRec(double z_src, double z_rec, double theta_src);
+  double calc_r(double theta_src,double s, double R);
+  double calcJ(double s, double theta_src, double d_theta, double R);
+  double calcPfromArcLength(double z_src, double z_rec, double J, double theta_src);
+  double calcTransmissionLoss(double theta_src, double z_src, double z_rec, double R, double d_theta);
 
 
 
   protected:
   //Inputs
-    float m_c_0, m_water_depth;
-    float m_time_int, m_gradient;
+    double m_c_0, m_water_depth;
+    double m_time_int, m_gradient;
